+++ examples/next-typescript-starter/app/register/firebase.ts
@@ -4,7 +4,7 @@ import {
   GoogleAuthProvider,
   signInWithPopup,
   signOut,
-  useDeviceLanguage
+  useDeviceLanguage as setDeviceLanguage
 } from 'firebase/auth';
 
 const CREDENTIAL_ALREADY_IN_USE_ERROR = 'auth/credential-already-in-use';
@@ -19,7 +19,7 @@ export const getGoogleProvider = (auth: Auth) => {
   const provider = new GoogleAuthProvider();
   provider.addScope('profile');
   provider.addScope('email');
-  useDeviceLanguage(auth);
+  setDeviceLanguage(auth);
   provider.setCustomParameters({
     display: 'popup'
   });