+++ src/auth/jwt/verify.ts
@@ -14,6 +14,7 @@ export interface VerifyOptions {
   currentDate?: Date;
   checkRevoked?: boolean;
   referer?: string;
+  experimental_enableTokenRefreshOnExpiredKidHeader?: boolean;
 }
 
 const keyMap: Map<string, KeyLike> = new Map();
@@ -61,7 +62,8 @@ export async function verify(
     }
     if (payload.nbf > currentTimestamp) {
       throw new errors.JWTExpired(
-        'jwt not active: ' + new Date(payload.nbf * 1000).toISOString()
+        'jwt not active: ' + new Date(payload.nbf * 1000).toISOString(),
+        payload
       );
     }
   }
@@ -73,7 +75,8 @@ export async function verify(
 
     if (currentTimestamp >= payload.exp) {
       throw new errors.JWTExpired(
-        'token expired: ' + new Date(payload.exp * 1000).toISOString()
+        'token expired: ' + new Date(payload.exp * 1000).toISOString(),
+        payload
       );
     }
   }